bind server connection state
