+++ Client/Rubberduck.Editor/App.xaml.cs
@@ -52,7 +52,6 @@ using System.IO.Abstractions;
 using System.Reflection;
 using System.Threading;
 using System.Windows;
-using ILogger = Microsoft.Extensions.Logging.ILogger;
 
 namespace Rubberduck.Editor
 {
@@ -108,16 +107,20 @@ namespace Rubberduck.Editor
 
                     splash.UpdateStatus("Initializing language server protocol (addin/editor)...");
                     await _editorServer.StartupAsync();
+
+                    if (!string.IsNullOrWhiteSpace(_options.WorkspaceRoot))
+                    {
+                        splash.UpdateStatus("Initializing language server protocol (editor/server)...");
+                        await _languageClient.StartupAsync();
+                    }
                 }
                 else if (_settings.Settings.LanguageClientSettings.RequireAddInHost)
                 {
                     throw new InvalidOperationException("Editor is not configured for standalone execution.");
                 }
 
-                splash.UpdateStatus("Initializing language server protocol (editor/server)...");
-                await _languageClient.StartupAsync();
-
                 splash.UpdateStatus("Quack!");
+
                 ShowEditor();
                 splash.Close();
             }
@@ -204,7 +207,9 @@ namespace Rubberduck.Editor
         private void ConfigureServices(IServiceCollection services)
         {
             services.AddSingleton<Func<ILanguageServer>>(provider => () => _editorServer.EditorServer);
-            services.AddSingleton<Func<ILanguageClient>>(provider => () => _languageClient.LanguageClient);
+            services.AddSingleton<Func<ILanguageClient?>>(provider => () => _languageClient.LanguageClient);
+            services.AddSingleton<LanguageClientApp>(provider => _languageClient);
+
             services.AddSingleton<ServerStartupOptions>(provider => _options);
             services.AddSingleton<Process>(provider => Process.GetProcessById((int)_options.ClientProcessId));
 
@@ -303,64 +308,6 @@ namespace Rubberduck.Editor
             _editorServer?.Dispose();
             _languageClient?.Dispose();
             _tokenSource.Dispose();
-        }
-    }
-
-    public class WorkspaceRootResolver
-    {
-        private readonly ILogger _logger;
-        private readonly ISettingsProvider<RubberduckSettings> _settings;
-
-        public WorkspaceRootResolver(ILogger logger, ISettingsProvider<RubberduckSettings> settings)
-        {
-            _logger = logger;
-            _settings = settings;
-        }
-
-        protected TraceLevel TraceLevel => _settings.Settings.LoggerSettings.TraceLevel.ToTraceLevel();
-
-        public Uri GetWorkspaceRootUri(ServerStartupOptions options)
-        {
-            var settings = _settings.Settings.LanguageClientSettings;
-            var setting = settings.WorkspaceSettings.GetSetting<DefaultWorkspaceRootSetting>();
-            var uri = setting.DefaultValue;
-
-            var argsRoot = options.WorkspaceRoot;
-
-            if (argsRoot is null && options.ClientProcessId == default && settings.RequireAddInHost)
-            {
-                _logger.LogDebug("An add-in host is required and should have provided a workspace root command-line argument. The current configuration does not support standalone execution.");
-                throw new NotSupportedException("An add-in host is required and should have provided a workspace root command-line argument. The current configuration does not support standalone execution.");
-            }
-            else if (argsRoot is null)
-            {
-                // editor is running standalone without an addin client connection.
-                _logger.LogDebug("No workspace root commad-line argument was specified, but configuration supports standalone execution. Using default workspace root; there is no project file or workspace folder yet.");
-                return setting.DefaultValue;
-            }
-
-            if (Uri.TryCreate(argsRoot, UriKind.Absolute, out var argsRootUri))
-            {
-                uri = argsRootUri;
-            }
-            else
-            {
-                _logger.LogWarning("Could not parse value '{argsRoot}' into a valid URI. Falling back to default workspace root.", argsRoot);
-            }
-
-            if (settings.WorkspaceSettings.RequireDefaultWorkspaceRootHost && !uri.LocalPath.StartsWith(setting.DefaultValue.LocalPath))
-            {
-                _logger.LogWarning(TraceLevel, $"Configuration requires a workspace root under the default folder, but a folder under a different root was supplied.", uri.LocalPath);
-                throw new NotSupportedException($"Configuration requires a workspace root under the default folder, but a folder under a different root was supplied.");
-            }
-
-            if (!settings.WorkspaceSettings.EnableUncWorkspaces && uri.IsUnc)
-            {
-                _logger.LogWarning(TraceLevel, $"UNC URI is not allowed: {nameof(settings.WorkspaceSettings.EnableUncWorkspaces)} setting is disabled. Default setting value will be used instead.", uri.ToString());
-                uri = setting.DefaultValue;
-            }
-
-            return uri;
         }
     }
 }
