+++ Client/Rubberduck.UI/Shell/StatusBar/ServerConnectionState.cs
@@ -3,6 +3,7 @@
     public enum ServerConnectionState
     {
         Disconnected,
+        Connecting,
         Connected
     }
 }
