+++ Shared/Rubberduck.Resources/Icons/RubberduckIcons.resx
@@ -118,6 +118,12 @@
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
   <assembly alias="System.Windows.Forms" name="System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" />
+  <data name="circle_check_solid" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\circle-check-solid.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
+  <data name="circle_exclamation_solid" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\circle-exclamation-solid.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
   <data name="file_code_regular" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\Resources\file-code-regular.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
   </data>
@@ -129,5 +135,8 @@
   </data>
   <data name="folder_closed_regular" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\Resources\folder-closed-regular.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
+  <data name="hourglass_half_solid" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\hourglass-half-solid.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
   </data>
 </root>
